CODIGO/requirements.txt
customtkinter
Pillow
